2270. Number of Ways to Split Array. Solved. Preparation with prefix sum, easy logic. Use long instead of int

+++ LeetCodeTasks/NumberOfWaysToSplitArray.cs
@@ -18,7 +18,7 @@ namespace LeetCodeTasks
             }
 
             // solution
-            int indexSum = 0;
+            long indexSum = 0;
             for (int i = 0; i < nums.Length - 1; ++i)
             {
                 indexSum = indexSum + nums[i];
